feat: add config entry to specify field types
A `register.<reg>.types.<field>` entry can be added to the configuration
file to specify which type to use to represent a field. The value is
converted using `TryFrom` then unwrapping. Due to the use of a trait,
the functions to access the field cannot be constant.

A possible workaround for this would be to allow to specify how to
convert the field in the configuration, e.g. by providing the path of a
`const` function.

Change-Id: Id37d1fce5c88d53cf6bf8abd09d8168a68102876
Signed-off-by: Ludovic Mermod <ludovic.mermod@arm.com>