feat: add subcommand to list potential enums
Splits the tool into two commands: `generate` (the previous default
behavior) and `enums`. The latter list register fields that may be
represented as Rust enums instead of raw integers.

It also exposes a `--generate-stubs` flag which outputs stub
implementations of the enums and configuration entries to use them and
a `--skip-existing` one to avoid displaying already implemented enums.

Change-Id: I568a9d382b0256131be3b8c6c671ecd5b0fe70a2
Signed-off-by: Ludovic Mermod <ludovic.mermod@arm.com>